float -> int

+++ flow.ml
@@ -20,17 +20,17 @@ let neighbours g id = (* a graph -> id -> id list  *)
 
 
 
-let flow_var g idlist =(*(float*float) graph ->  id list -> float *)
+let flow_var g idlist =(*(int*int) graph ->  id list -> int *)
   let inner acu l = match l with
     |[] -> failwith "not possible"
     |[x]-> acu
     |x::(y::rest) -> (match List.find(fun (a,_) -> a=y) (Graph.out_arcs g x) with
       |_,(a,b) -> if (b-a) < acu then inner (b-a) (y::rest) else inner acu (y::rest) ) 
   in
-  inner Float.infinity(*and beyond*) idlist ;;
+  inner int.infinity(*and beyond*) idlist ;;
 
 
-let variation_graph gr = (*(float*float) graph -> float graph *) 
+let variation_graph gr = (*(int*int) graph -> int graph *) 
   let g2 = Graph.e_fold gr (fun g id1 id2 (a,b) -> Graph.new_arc g id1 id2 (b-a)) (clone_nodes gr) in
   Graph.e_fold g2 (fun g id1 id2 (a,b) -> Graph.new_arc g id2 id1 a) g2
 ;;
